moving non-dist tests to new directory structure
