+++ tests/unit/alexnet_model.py
@@ -1,29 +1,10 @@
-import copy
-
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-import deepspeed.comm as dist
-
-import pytest
-
 import deepspeed
+import deepspeed.comm as dist
 import deepspeed.runtime.utils as ds_utils
-
-from deepspeed.runtime.pipe.topology import PipeDataParallelTopology
-
-PipeTopo = PipeDataParallelTopology
 from deepspeed.runtime.pipe.module import PipelineModule, LayerSpec
-
-from .common import distributed_test
-
-
-def rel_diff(A, B):
-    return abs(A - B) / abs(A)
-
-
-# All models
-from .simple_model import args_from_dict
 
 
 class AlexNet(nn.Module):
@@ -137,22 +118,22 @@ def cifar_trainset(fp16=False):
 
 
 def train_cifar(model,
-                config_dict,
+                config,
                 num_steps=400,
                 average_dp_losses=True,
                 fp16=True,
                 seed=123):
-    with torch.random.fork_rng(devices=[torch.cuda.current_device()]):
+    with torch.random.fork_rng(devices=[torch.cuda.current_device()], enabled=False):
         ds_utils.set_random_seed(seed)
 
         # disable dropout
         model.eval()
 
         trainset = cifar_trainset(fp16=fp16)
-        config_dict['local_rank'] = dist.get_rank()
+        config['local_rank'] = dist.get_rank()
 
         engine, _, _, _ = deepspeed.initialize(
-            config=config_dict,
+            config=config,
             model=model,
             model_parameters=[p for p in model.parameters()],
             training_data=trainset)
@@ -171,102 +152,3 @@ def train_cifar(model,
             losses = loss_tensor.tolist()
 
     return losses
-
-
-@pytest.mark.skip(reason="been seeing nondeterministic failures, skipping for now")
-@pytest.mark.parametrize('topo',
-                         [
-                             PipeTopo(num_pp=1,
-                                      num_dp=4),
-                             PipeTopo(num_pp=2,
-                                      num_dp=2),
-                             PipeTopo(num_pp=4,
-                                      num_dp=1),
-                         ])
-def test_pipe_cifar10(topo, tmpdir):
-    config_dict = {
-        "train_batch_size": 16,
-        "train_micro_batch_size_per_gpu": 4,
-        "steps_per_print": 20,
-        "optimizer": {
-            "type": "Adam",
-            "params": {
-                "lr": 0.001,
-                "betas": [0.9,
-                          0.999],
-                "eps": 1e-8,
-                "weight_decay": 3e-7
-            }
-        },
-        "zero_optimization": {
-            "stage": 0
-        },
-        "fp16": {
-            "enabled": False
-        },
-        "pipeline": {
-            "seed_layers": True,
-            "activation_checkpoint_interval": 1
-        }
-    }
-    args = args_from_dict(tmpdir, config_dict)
-
-    # Allocate model for consistent initial weights.
-    init_net = AlexNetPipe()
-
-    @distributed_test(world_size=4)
-    def _helper(topo, tmpdir, steps=500):
-        assert steps >= 100
-
-        base_net = copy.deepcopy(init_net)
-        base_model = PipelineModule(layers=base_net.to_layers(),
-                                    num_stages=1,
-                                    loss_fn=nn.CrossEntropyLoss())
-
-        # Train with just data parallelism
-        base_losses = train_cifar(base_model,
-                                  args,
-                                  num_steps=steps,
-                                  fp16=config_dict['fp16']['enabled'])
-
-        test_net = copy.deepcopy(init_net)
-        test_model = PipelineModule(layers=test_net.to_layers(),
-                                    topology=topo,
-                                    loss_fn=nn.CrossEntropyLoss())
-
-        #test_model = AlexNetPipe(num_classes=10,
-        #                         topology=test_topo,
-        #                         seed_layers=config_dict['pipeline']['seed_layers'])
-        test_losses = train_cifar(test_model,
-                                  args,
-                                  num_steps=steps,
-                                  fp16=config_dict['fp16']['enabled'])
-
-        abs_diffs = [l0 - l1 for l0, l1 in zip(base_losses, test_losses)]
-        rel_diffs = [rel_diff(l0, l1) for l0, l1 in zip(base_losses, test_losses)]
-        if dist.get_rank() == 0:
-            print(
-                f'abs min={min(abs_diffs)} max={max(abs_diffs)} avg={sum(abs_diffs)/len(abs_diffs)}'
-            )
-            print(
-                f'rel min={min(rel_diffs)} max={max(rel_diffs)} avg={sum(rel_diffs)/len(rel_diffs)}'
-            )
-            print(
-                f'first: base={base_losses[0]} test={test_losses[0]} abs={abs_diffs[0]} rel={rel_diffs[0]}'
-            )
-
-            for lastX in [1, 10, 100]:
-                base_avg = sum(base_losses[-lastX:]) / lastX
-                test_avg = sum(test_losses[-lastX:]) / lastX
-                print(
-                    f'last-{lastX}: base={base_avg} test={test_avg} abs={base_avg - test_avg} rel={rel_diff(base_avg, test_avg)}'
-                )
-
-        lastX = 100
-        base = base_losses[-lastX:]
-        base_avg = sum(base) / len(base)
-        test = test_losses[-lastX:]
-        test_avg = sum(test) / len(test)
-        assert rel_diff(base_avg, test_avg) < 0.03
-
-    _helper(topo, tmpdir)
